Refactored CachingService
