+++ PokemonLookup/PokemonLookup.Infrastructure/InfrastructureServiceExtensions.cs
@@ -1,6 +1,7 @@
 using Microsoft.Extensions.DependencyInjection;
 using PokemonLookup.Application;
 using PokemonLookup.Application.Services;
+using PokemonLookup.Infrastructure.Caching;
 using PokemonLookup.Infrastructure.ExternalLookup;
 
 namespace PokemonLookup.Infrastructure;
@@ -18,7 +19,7 @@ public static class InfrastructureServiceExtensions
     public static IServiceCollection AddInfrastructureServices(this IServiceCollection services)
     {
         services.AddSingleton<IInputChecker, InputChecker>();
-        services.AddScoped<ICachingService, CachingService>();
+        services.AddScoped<ICachingService, DatabaseCachingService>();
         services.AddSingleton<IApiRequester, ApiRequester>();
         services.AddSingleton<IPokemonApiRequester, PokemonApiRequester>();
         services.AddScoped<IPokemonLibrary, PokemonLibrary>();
